The problem is Proof of Stake blocks are not stored when the previous transaction of the CoinStake is missing. So the initial download process is broken.
To solve this:
The reference block is now the best block, whatever the proof.

The client asks for missing blocks when it receives a PoS block and
it doesn't have the CoinStake previous transaction.

Block request are retried immediatly during initial download because
the same last block is asked multiple times until we have the full
chain.

Solution inspired by Sigmike.